+++ api/operator/v1beta1/zz_generated.deepcopy.go
@@ -3729,6 +3729,11 @@ func (in *VMAgentSpec) DeepCopyInto(out *VMAgentSpec) {
 		*out = new(ManagedObjectsMetadata)
 		(*in).DeepCopyInto(*out)
 	}
+	if in.APIServerConfigDeprecated != nil {
+		in, out := &in.APIServerConfigDeprecated, &out.APIServerConfigDeprecated
+		*out = new(APIServerConfig)
+		(*in).DeepCopyInto(*out)
+	}
 	if in.APIServerConfig != nil {
 		in, out := &in.APIServerConfig, &out.APIServerConfig
 		*out = new(APIServerConfig)